feat: add global scroll-to-top component and fix Next.js dependencies
