+++ next.config.mjs
@@ -30,5 +30,5 @@ const nextConfig = {
   },
 };
 
-export default withPWA(nextConfig);
+export default nextConfig;
